Started work on project
Code Worked on 4/3/19. Renamed files to their proper names. Linked the javascript file into the html file. Added the function to define the data array and populate it. Added the function to write the table data from the array to the table in the html. Type the beginning of the definecolumns function.